Be explicit about which path we bundle into and cache it explicitly
Because this is a language: nodejs project, Travis doesn't automatically
infer that bundler stuff should be cached, or automatically append
caching parameters to the bundler invocations. This adds manual caching
for bundler.

Note: the default path for bundler caching on travis is vendor/bundle,
but this opts to use .vendor/bundle so that the files are ignored by
most tooling, like tslint and webpack configurators. It seems better to
ignore the whole folder using filesystem conventions instead of adding
exceptions to each tool looking at the filesystem.